less bonkers version strings
